Set a blank excerpt for all pages
This commit sets a blank Jekyll excerpt as the default for all pages
([docs on excerpts][1]). Without this configuration, Jekyll tries to
generate a page excerpt using the first paragraph of each page. However,
we do not use the page excerpt anywhere in this project (since we don't
have an equivalent of a "list of all blog posts" page). Setting a blank
excerpt decreases website build time.

More importantly, in a later commit I plan to wrap HTML generated from a
Jupyter notebook in a `{% raw %}` tag ([docs][2]). Without this tag,
complicated LaTeX will break Jekyll builds. However, when Jekyll tries
to generate page excerpts, it breaks because the excerpt contains the
`{% raw %}` tag but not the `{% endraw %}` tag.

[1]: https://jekyllrb.com/docs/posts/#post-excerpts
[2]: https://shopify.github.io/liquid/tags/raw/