Wrap `jekyll-raw` cells with {% raw %}
In LaTeX and Python it is not uncommon to have Jekyll/Liquid keywords,
like the substrings `{{` or `}}`. To tell Jekyll not to process these
as Jekyll commands, users can give cells the tag `jekyll-raw`. The
generated HTML will be wrapped in a `{% raw %}` tag.

To test this, I did the following:

1. Added a Python cell to
   `./jupyter_book/book_template/content/features/notebooks.ipynb` with
   the code `'{{{hello}}}'`.
2. Checked that `make book build` fails.
3. Added the `jekyll-raw` tag to the cell.
4. Checked that `make book build` succeeds, and that
   http://127.0.0.1:4000/jupyter-book/features/notebooks.html has the
   Python code `'{{{hello}}}'`.