Suggested update to description of toc.yml in Anatomy of a Jupyter Book.
The current description of the toc.yml is out of sync with the actual toc.yml
file, which might be confusing for users reading the docs.
Replaced the code snippet with the sample from the current toc.yml.
Also added a few sentences to denote how the titles of the documents
is determined, as that has changed since the previous toc example.
Includes link to relevant page in the beta features docs