Modify toc depth to show links to appropriate sections.
The pdf-building docs use a \{content\} directive to list section
headings within the document to highlight two different ways of
building pdfs. Change depth from (1) to (2) so that the relevant
headings are shown.

Alternatively, it might be better to replace the content directive
entirely as it adds a title and the doc title, neither of which is
necessary in this case.