ENH: Modify page to exec nbs by default.
Currently, the page command has a CL option that accepts one of the
four execution kwargs (off, cache, force, or auto) to determine
whether a page should be executed during conversion.

This change replaces the existing execute flag with a binary
flag (--execute/--no-execute) that does not need the keyword
options to be passed in.

Also changes the default execution behavior to True, ignoring
the book-level configuration for notebook execution.