🧪 TESTS: Add Windows OS to CI
This involved numerous fixes to the code and tests, to use posix strings.
There was also an issue of the tempdir being created on a different drive.
Therefore, a pytest cmdline option was added to set the location of this directory.
Additionally, pip caching was set up on the GitHub CI, and and pytest-timeout config

Co-authored-by: phaustin <https://github.com/phaustin>
Co-authored-by: foster999 <https://github.com/foster999>"