Fix header reference example
Apologies in advance for consuming review cycles for such a minor fix. I thought I misunderstood the documentation and figured I'd send the correction in case others get confused. In the doc, the section being referenced is: `Reference section labels`  but the documentation names the section as `Cross-references and labels`.