Make punctuation on "If using" lines consistent
VERY MINOR REVISION:
I changed the two "If using..." lines to end in colons. Previously, one ended in on punctuation, and one ended in a period.  I just thought it would be tidier if they had consistent punctuation. 

Feel free to commit with or without this suggested revision, as you see fit. Thanks, @choldgraf!