maint: remove precautionary upper version bounds on dependencies
For all dependencies with a version upper bound that is clearly
precautionary, where the upper bound is not constraining the version
currently - remove the upper bound.

For all remaining upper bounds, add a fixme note to clarify that we need
to understand if its just a precautionary bound, or if we can remove it
as a switch to a new principle where we stop adding upper bounds on
dependencies.