Update banner to use github issue (#2282)
This updates our banner to use a GitHub issue for ongoing information on
the 2.0 switch (instead of using a blog post). My reasoning is that we
can use that issue as the "source of truth" for external people that
want to learn more about this transition, what to expect, where to learn
more, etc. This might change over time so better to use a github issue
than a blog post.

Here's the issue:

- https://github.com/jupyter-book/jupyter-book/issues/2281

Any objections from the @jupyter-book/core-team about this? I'm open to
alternative suggestions!