Update new issues template to point to discussions forum (#2292)
This adds a link to the Jupyter Book discussion forum for "new issue"
templates.

I also propose that we update our "bug report" issue to just be a
general bug report, not something "specific" to Jupyter Book 2. However
we should make clear in the issue template that users should specify if
this relates to JB 1 or 2.

This way, the new issue button menu should look like this:

1. Open a bug report
2. Open a blank issue
3. Ask a question (links to forum)
4. Request an enhancement (links to forum)

Over time, if people are still opening unhelpful blank issues and not
using the forum, then we can consider removing that option so only
maintainers can open blank issues.