MAINT: prepare for new release (#2327)
We need to make a path release in order to update our deployed
documentation. We also need to address the sphinx-book-theme breakage
_and_ simplify our maintenance burden.

This PR tightens our pins, treating JB 1 as an [app rather than a
library](https://iscinumpy.dev/post/app-vs-library/#application). It
also updates to Hatch and playwright instead of pyppeteer (which has
been deprecated for a while).