MAINT: adding cooldown for dependabot (#2470)
Seeing #2469 reminded me that we talked with @stefanv about adding a
cooldown for updates to be at least somewhat safeguarded about malicious
releases.

Using 2 weeks feels like a good middle ground, but I'm not married to
the actual number (IMO it should not be shorter -- if any version is
critical enough to require ASAP updates, that can always be done
manually by a maintainer).

If this is a thumbs up, I'm happy to take care of opening PRs to
all/most of our other repos.