📖 Remove references "Jupyter Book 2" (#2496)
This cleans up references to "Jupyter Book 2" from the install guide. As
there is a full release now, you don't need to specify a version to get
an alpha/beta release.

Also, changing the language because Jupyter Book 2 _is_ Jupyter Book.

I've left references to Jupyter Book 2 in the history and upgrade
sections because in those places we are talking about the differences
between v1 and v2.

Co-authored-by: Rowan Cockett <rowanc1@gmail.com>