Add a table for our team priorities roadmap and use the github-issue-tables plugin (#2490)
This PR does two related things:

1. It adds a table version of our team priorities roadmap, so that this
is easier to publicly find and navigate.
2. It replaces the custom github issue tables logic with a myst plugin
that I put up in the myst-plugins repo here:
https://github.com/jupyter-book/myst-plugins/tree/main/plugins/github-issue-table

That plugin gives us a directive we can re-use so that we aren't writing
the same custom issue table code everywhere, and so we can improve that
code in one place to re-use it here and elsewhere.

Here's the voting board:

[Live
link](https://jupyter-book--2490.org.readthedocs.build/2490/community/vote/)

````
```{issue-table} org:jupyter-book is:open is:issue
:columns: title, repo, last_updated, reactions
:sort: reactions-desc
:limit: 30
:date-format: relative
```
````

Here's the project board for the roadmap:

[Live
link](https://jupyter-book--2490.org.readthedocs.build/2490/community/roadmap/)

````
```{issue-table} https://github.com/orgs/jupyter-book/projects/1
:columns: title, Team Priority, reactions, updated
:sort:Team Priority-asc,reactions-desc
:limit:1000
:date-format: relative
```
````

I think I triggered a bug with the PR so documented it here:

- https://github.com/jupyter-book/myst-plugins/issues/13

This PR includes a workaround where we download the plugin locally as
part of our build step until it's resolved.

---------

Co-authored-by: Angus Hollands <goosey15@gmail.com>